stm32: fix fmc-related build failures on some F4's
